+++ plugins/newspack-plugin/includes/plugins/woocommerce/class-woocommerce-cover-fees.php
@@ -238,8 +238,9 @@ class WooCommerce_Cover_Fees {
 		$total = self::get_total_with_fee( $subtotal );
 		if ( ! function_exists( 'wc_price' ) ) {
 			$donation_settings = Donations::get_donation_settings();
+			$currency_symbol   = is_wp_error( $donation_settings ) ? '$' : $donation_settings['currencySymbol'];
 			// Just one decimal place, please.
-			return $donation_settings['currencySymbol'] . ( (float) number_format( $total - $subtotal, 2 ) );
+			return $currency_symbol . ( (float) number_format( $total - $subtotal, 2 ) );
 		}
 		return \wc_price( $total - $subtotal );
 	}